Commit this mess
